LeftBox 완성
