+++ react-weato/src/components/sections/Login.js
@@ -1,12 +1,10 @@
 import React, {useState} from 'react'
 import {Link} from 'react-router-dom';
 import Section from '../../HOC/section';
-import "../../static/style.css";
 import { faUser } from "@fortawesome/free-solid-svg-icons";
 import {faLock} from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import bgImage from '../../image/back.png';
-
+// import bgImage from "../../image/back.png";
 function Login() {
   const [ID, setID] = useState("");
   const [PW, setPW] = useState(""); 
@@ -25,18 +23,20 @@ function Login() {
 
   return (
     <Section id='Login'>
-      <div
+      {/* <div
           className='home-content p-5'
           style={{ backgroundImage: `url(${bgImage})` }}
-      >
+      >  */}
       <div className='container'>
         <div className='section-header pt-5 pb-5 text-center'>
           <div className="main">
             <div className="sub-main">
               <h1 className="Login-title">Login</h1>
               <form>
-                <FontAwesomeIcon icon={faUser} className="ID-icon"/>
-                <FontAwesomeIcon icon={faLock} className="PW-icon"/> 
+                <div className="login-icons">
+                  <FontAwesomeIcon icon={faUser} className="ID-icon"/>
+                  <FontAwesomeIcon icon={faLock} className="PW-icon"/> 
+                </div>
                 <input name="input_ID" type="text" placeholder="ID" value={ID} onChange={onIDHandler} className="login-input"></input>                
                 <input name="input_PW" type="text" placeholder="PW" value={PW} onChange={onPWHandler}className="login-input"></input>
                 <button type="submit" onSubmit={onSubmit} className="login-button">Login</button>
@@ -46,7 +46,7 @@ function Login() {
           </div>
         </div>
       </div>
-      </div>
+      {/* </div> */}
     </Section>
     );
   }
