+++ react-weato/src/static/css/LeftBox.css
@@ -1,20 +1,145 @@
 .leftBox{
+    position: relative;
     display: grid;
     grid-template-columns: 100%;/* 각 행(세로줄)의 길이 */
     grid-template-rows: 50px 220px 130px 250px;/* 각 열(가로줄)의 길이 */
     gap: 10px; /* 자식요소간의 간격 */
+    color: #4A4A4A;
 }
 .leftBox-layout{
     width: 100%;
     height:100%;
-    border: 2px solid black;
+}
+.locate-icon{
+    height:24px;
+    width:18px;
+    position: absolute;
+    margin-left: 12%;
+    margin-top:2%
+}
+.search-button{
+    margin-top: 1.5%;
+    height:33px;
+    width: 33px;
+    background: url( "../../../public/img/icon_search_submit.png" );
+    background-size: cover;
+    border:none;
+    cursor: pointer;
+    position: absolute;
+    margin-left: 1%;
+    border-radius: 4px;
+    background-color: aliceblue;
+}
+.search-button:hover{
+    border:2px solid #99d8f3;
+    box-shadow: 2px 2px 0 rgba(147, 210, 241, 0.5);
 }
 .leftBox-layout>input{
     margin-left: 10%;
     margin-top: 1%;
     height: 80%;
-    width:80%;
+    width:70%;
     padding-left: 7%;
     font-size: 10px;
     border-radius: 10px;
+    border:none;
 }
+.weather-content{
+    display: grid;
+    grid-template-columns: 50% 50%;/* 각 행(세로줄)의 길이 */
+    grid-template-rows: 220px;/* 각 열(가로줄)의 길이 */
+    gap: 10px; /* 자식요소간의 간격 */
+}
+.weather-info{
+    height: 100%;
+    width:100%;
+    padding:3%;
+}
+.weather-info .today{
+    margin-top:2%;
+    font-weight: 500;
+    font-size: 12px;
+}
+.weather-info .area{
+    margin-top: 1%;
+    font-weight:600;
+    font-size:18px;
+}
+.weather-info .avg-temp{
+    font-size :45px;
+    font-weight:550;
+}
+.weather-info .avg-rainy,.avg-windy, .avg-hum{
+    display: flex;
+    margin-left: 1%;
+    gap:10%;
+    font-weight:800;
+    font-size: 14px;
+    margin-top:1%;
+}
+.rainy-icon,.windy-icon,.hum-icon{
+    width:55px;
+    height:23px;
+    font-size:15px;
+    background-color: white;
+    padding-left: 4%;
+    border-radius: 40px;
+}
+
+.four-weather{
+    display: grid;
+    grid-template-columns: 46.5% 46.5%;
+    grid-template-rows: 60px 60px;
+    margin-left: 7%;
+    gap: 8px;
+}
+.four-weather-layout{
+    width:100%;
+    height:100%;
+}
+.TMA{
+    background-color: #c35353;
+}
+.Acu{
+    background-color: #DE59A9;
+}
+.channel{
+    background-color: #8572CE;
+}
+.news{
+    background-color: #74B472;
+}
+
+.TMA,.Acu,.channel,.news {
+    display: grid;
+    grid-template-columns: 65% 30%;
+    grid-template-rows: 60px;
+    gap: 8px;
+    padding-left: 5%;
+    color:white;
+    border-radius: 15px;
+}
+.TMA-layout,.Acu-layout,.channel-layout,.news-layout{
+    width:100%;
+    height:100%;
+}
+.TMA-name,.Acu-name,.channel-name,.news-name{
+    padding-top: 2%;
+    font-size:12px;
+    height:10px;
+}
+.TMA-temp,.Acu-temp,.channel-temp,.news-temp{
+    font-size:25px;
+    font-weight: 500;
+}
+.TMA-avg-rainy,.TMA-avg-windy,.TMA-avg-hum,.Acu-avg-rainy,.Acu-avg-windy,.Acu-avg-hum,.channel-avg-rainy,.channel-avg-windy,.channel-avg-hum,.news-avg-rainy,.news-avg-windy,.news-avg-hum{
+    display: flex;
+    margin-left: 1%;
+    gap:8%;
+    font-weight:800;
+    font-size: 10px;
+    padding-top: 4%;
+}
+
+
+
