+++ react-weato/src/static/css/WeatoAdd.css
@@ -1,0 +1,248 @@
+.add-main{/* 등록페이지 배경 */
+    justify-content: center;
+    display: grid; 
+    background:url(../../../public/img/bg_cloud.png);
+    background-size: cover;
+    padding:8%;
+    /* opacity: 0.5; */
+} 
+
+.add-Container{/* 등록 페이지 창 */
+    justify-content: center;
+    height:680px;
+    width: 500px;
+    box-shadow: 11px 12px 13px 12px rgb(207,207,207);
+    padding:8%;
+    border-radius: 50px;
+    background-color: white;
+}
+
+.add-content-box{/* 페이지 레이아웃 */
+    margin-top: 5%;
+    display: grid;
+    grid-template-columns: 180px 230px; /* 각 행(세로줄)의 길이 */
+    grid-template-rows: 400px; /* 각 열(가로줄)의 길이 */
+    gap: 10px; /* 자식요소간의 간격 */
+}
+.add-box{  
+    width: 100%;
+    height: 100%;
+    font-size:10px;
+}
+
+.add-title{/* 페이지 제목 */
+    font-size:10px;
+}
+
+.title-display{
+    display: flex;
+    margin-top: 1%;
+}
+
+.locate-Img{/* 위치 이미지 */
+    height: 30px;
+    width:20px;
+    background: url(../../../public/img/icon_locate_current_location.png);
+    background-size: cover;
+    border:none;
+    cursor: pointer;
+}
+
+.area-name{/* 현재 지역 이름 */
+    margin-left: 3%;
+    font-size:20px;
+    font-weight:700;
+    color:red;
+    width:60%;
+}
+.gps-button{/* gps 버튼 */
+    margin-top: 2%;
+    height:15px;
+    width: 15px;
+    background: url( "../../../public/img/icon_button_location.png" );
+    background-size: cover;
+    border:none;
+    cursor: pointer;
+}
+.area-link{/* 지역 찾기 링크 */
+    font-size: 9px;
+    height:3px;
+}
+.add-today{/* 오늘 날짜/시간 */
+    margin-left:75%;
+    font-size:10px;
+}
+.add-title-box>hr{ /*구분선*/
+    margin:0px;
+}
+/* 오른쪽 레이아웃 */
+.add-box-right{
+    display:grid;
+    grid-template-columns: 230px; /* 각 행(세로줄)의 길이 */
+    grid-template-rows: 80px 80px 80px 80px 40px; /* 각 열(가로줄)의 길이 */
+    gap: 8px; /* 자식요소간의 간격 */
+}
+.add-box-right-rayout{
+    width: 100%;
+    height: 100%;
+}
+.add-box-right-title{
+    font-weight: 500;
+}
+
+/* 체크박스 레이아웃 */
+.add-box-right-state{
+    display:flex;
+    margin-left:20%;
+    margin-top: 3%;
+    gap: 30%;
+}
+/* 체크박스 스타일*/
+.add-box-right-state>input+label{
+    height: 10px;
+    width:10px;
+    border-radius: 50%;
+    display: inline-block;
+    cursor: pointer;
+    box-sizing: border-box;
+    border: 2px solid #bcbcbc; 
+}
+/* 체크박스 체크 시 활성화 */
+.add-box-right-state>input:checked+label{
+	background-color:#5BB7DE;
+}
+ 
+.add-box-right-state>input{
+    display:none;
+}
+/* 아이콘 이미지 */
+.add-box-right-img{
+    display:flex;
+    margin-left:18%;
+    margin-top: 2%;
+    gap: 24%;
+}
+.add-box-right-img>img{
+    width:20px;
+    height:20px;
+}
+/* 아이콘 설명 */
+.add-box-right-img-label{
+    display:flex;
+    margin-left:18%;
+    margin-top: 1%;
+    gap: 25%;
+}
+/* 구름 상태 체크박스 */
+.add-box-right-state-sky>input+label{
+    height: 10px;
+    width:10px;
+    border-radius: 50%;
+    display: inline-block;
+    cursor: pointer;
+    box-sizing: border-box;
+    border: 2px solid #bcbcbc; 
+}
+.add-box-right-state-sky>input:checked+label{
+	background-color:#5BB7DE;
+}
+.add-box-right-state-sky>input{
+    display:none;
+}
+/* 구름 상태 레이아웃 */
+.add-box-right-state-sky{
+    display:flex;
+    margin-left:15%;
+    margin-top: 1%;
+    gap: 25%;
+}
+/* 구름 상태 아이콘 */
+.add-box-right-img-sky{
+    display:flex;
+    margin-left:12.8%;
+    margin-top: 1%;
+    gap: 19%;
+}
+.add-box-right-img-sky>img{
+    height: 20px;
+    width:20px;
+}
+/*구름 상태 아이콘 설명 */
+.add-box-right-img-label-sky{
+    display:flex;
+    margin-left:13%;
+    margin-top: 1%;
+    gap: 20%;
+}
+/* 비, 눈 상태 */
+.add-box-right-rainOrsnow{
+    display:flex;
+    gap: 8%;
+    margin-top:3%;
+    margin-left: 20%;
+}
+/* 비,눈 체크박스 */
+.add-box-right-rainOrsnow>input+label{
+    height: 10px;
+    width:10px;
+    border-radius: 50%;
+    display: inline-block;
+    cursor: pointer;
+    box-sizing: border-box;
+    border: 2px solid #bcbcbc; 
+}
+.add-box-right-rainOrsnow>input:checked + label{
+	background-color:#5BB7DE;
+}
+.add-box-right-rainOrsnow>input{
+    display:none;
+}
+.add-box-right-rainOrsnow>img{
+    height:30px;
+    width:30px;
+}
+
+/* 등록 페이지 레이아웃 왼쪽 */
+.imgInput{
+    margin-top: 10%;
+    margin-left: 20%;
+}
+.sampleImg{
+    height:100%;
+    width:100%;
+}
+
+.comment-input{/* 코멘트 입력 */
+    resize: none;
+    width:400px;
+    height:60px;
+    margin-top:3%;
+    margin-left: 3%;
+    border-radius: 40px;
+    padding-left: 5%;
+    font-size: 10px;
+    outline-color: #5BB7DE;
+}
+.comment-add-button{/* 코멘트 등록 버튼 */
+    width:55px;
+    height:20px;
+    font-size:10px;
+    border-radius: 40px;
+    margin-left: 83%;
+    background-color: #5BB7DE;
+    color: white;
+    border:none;
+}
+
+.comment-add:hover{/* 회원가입 버튼에 마우스 올릴 시 */
+    background-color:black;
+    color:white;
+    box-shadow: 2px 2px 0 rgb(0,0,0,0.2);
+}
+
+.NosampleImg{/* 사진 첨부 전 이미지 */
+    text-align:center;
+    width:100%;
+    height:200px;
+    background-color:#F2F2F2;
+}
