+++ react-weato/src/static/css/signUp.css
@@ -1,0 +1,33 @@
+.SignUp-title{/* 회원가입 제목 */
+    text-align:center;
+    font-size: 40px;
+    margin-bottom: 10%;
+    color:rgba(53, 53, 73, 0.847);
+}
+.SignUp-input{/* 회원가입 입력창 */
+    padding-left: 5%;
+    margin-top:5%;
+    width:300px;
+    height:45px;
+    border-radius: 60px;
+    border:2px solid darkgrey;
+    outline-color: #5BB7DE;
+    font-size:10px;
+    /* box-shadow: insert 0px 0px 25px 0px #888; */
+    
+}
+.SignUp-button{/* 회원가입 버튼 */
+    width:130px;
+    height:40px;
+    margin-top:10%;
+    border-radius: 60px;
+    background-color: #5BB7DE;
+    color:white;
+    font-size: 15px;
+    border:none;
+}
+.SignUp-button:hover{/* 회원가입 버튼에 마우스 올릴 시 */
+    background-color:black;
+    color:white;
+    box-shadow: 5px 5px 0 rgb(0,0,0,0.5);
+}
